Migrate non event related function in new helpers.py (see #84)
